manufacturing-ai-platform/services/gateway/requirements.txt
fastapi
uvicorn
requests
lxml
pycryptodome